only allow owner to edit/delete the room

+++ app/Http/Controllers/Api/RoomController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\Api;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
 use App\Http\Controllers\Controller;
 use App\Repositories\RoomRepository;
 use App\Repositories\CreditHistoryRepository;
@@ -38,7 +40,7 @@ class RoomController extends Controller
     public function add(CreateRoomRequest $request) {
         $data = $request->all();
         $data['available']  = true;
-        $data['owner_id']   = 1;
+        $data['owner_id']   = Auth::user()->id;
 
         $room = $this->roomRepo->insert($data, true);
         return response()->json($room);
@@ -50,6 +52,11 @@ class RoomController extends Controller
             throw new \App\Exceptions\DataNotFoundRestException('Room not found');
         } 
 
+        $belongTo = $object->owner()->id;
+        if ($belongTo != Auth::user()->id) {
+            throw new \App\Exceptions\GeneralRestException('WRONG_ROOM','You are not the owner of this room', 403); // forbiden
+        }
+
         $room = $this->roomRepo->update($id, $request->all());
         return response()->json($room);
     }
@@ -59,6 +66,11 @@ class RoomController extends Controller
         if ($object == null) {
             throw new \App\Exceptions\DataNotFoundRestException('Room not found');
         } 
+
+        $belongTo = $object->owner()->id;
+        if ($belongTo != Auth::user()->id) {
+            throw new \App\Exceptions\GeneralRestException('WRONG_ROOM','You are not the owner of this room', 403); // forbiden
+        }
 
         $status = $this->roomRepo->delete($id);
         if ($status) {
@@ -79,16 +91,16 @@ class RoomController extends Controller
             throw new \App\Exceptions\DataNotFoundRestException('Room not found');
         } 
 
-        $user = $this->userRepo->get(1);
+        $user = Auth::user();
 
         // check credit limit
-        $currentCredit = $this->creditHistRepo->checkCredit(1);
+        $currentCredit = $this->creditHistRepo->checkCredit($user->id);
         if ($currentCredit >= 40 && $user->type == 'PREMIUM') {
             throw new \App\Exceptions\CreditLimitRestException('Your account has been reached the monthly credit limit');
         } else if ($currentCredit >= 20 && $user->type == 'REGULER') {
             throw new \App\Exceptions\CreditLimitRestException('Your account has been reached the monthly credit limit, upgrade to PREMIUM account to get more limit.');
         }  else {
-            $creditHist = $this->creditHistRepo->insertHistory(1, 1); // add new history
+            $creditHist = $this->creditHistRepo->insertHistory($user->id, $object->id); // add new history
         }
 
         return response()->json(['available' => $object->available]);
